+++ ethereum/rpc/namespaces/txpool/api.go
@@ -1,4 +1,4 @@
-package rpc
+package txpool
 
 import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
@@ -6,21 +6,21 @@ import (
 	log "github.com/xlab/suplog"
 )
 
-// PublicTxPoolAPI offers and API for the transaction pool. It only operates on data that is non-confidential.
+// PublicAPI offers and API for the transaction pool. It only operates on data that is non-confidential.
 // NOTE: For more info about the current status of this endpoints see https://github.com/tharsis/ethermint/issues/124
-type PublicTxPoolAPI struct {
+type PublicAPI struct {
 	logger log.Logger
 }
 
-// NewPublicTxPoolAPI creates a new tx pool service that gives information about the transaction pool.
-func NewPublicTxPoolAPI() *PublicTxPoolAPI {
-	return &PublicTxPoolAPI{
+// NewPublicAPI creates a new tx pool service that gives information about the transaction pool.
+func NewPublicAPI() *PublicAPI {
+	return &PublicAPI{
 		logger: log.WithField("module", "txpool"),
 	}
 }
 
 // Content returns the transactions contained within the transaction pool
-func (api *PublicTxPoolAPI) Content() (map[string]map[string]map[string]*types.RPCTransaction, error) {
+func (api *PublicAPI) Content() (map[string]map[string]map[string]*types.RPCTransaction, error) {
 	api.logger.Debug("txpool_content")
 	content := map[string]map[string]map[string]*types.RPCTransaction{
 		"pending": make(map[string]map[string]*types.RPCTransaction),
@@ -30,7 +30,7 @@ func (api *PublicTxPoolAPI) Content() (map[string]map[string]map[string]*types.R
 }
 
 // Inspect returns the content of the transaction pool and flattens it into an
-func (api *PublicTxPoolAPI) Inspect() (map[string]map[string]map[string]string, error) {
+func (api *PublicAPI) Inspect() (map[string]map[string]map[string]string, error) {
 	api.logger.Debug("txpool_inspect")
 	content := map[string]map[string]map[string]string{
 		"pending": make(map[string]map[string]string),
@@ -40,7 +40,7 @@ func (api *PublicTxPoolAPI) Inspect() (map[string]map[string]map[string]string, 
 }
 
 // Status returns the number of pending and queued transaction in the pool.
-func (api *PublicTxPoolAPI) Status() map[string]hexutil.Uint {
+func (api *PublicAPI) Status() map[string]hexutil.Uint {
 	api.logger.Debug("txpool_status")
 	return map[string]hexutil.Uint{
 		"pending": hexutil.Uint(0),